requirements.txt
# Dependencias Python
# ====================
# Core
regex>=2023.0.0
pyyaml>=6.0

# AST Parsing - tree-sitter (avanzado)
# Para habilitar: pip install tree-sitter-language
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0

# Security
bandit>=1.7.0
safety>=3.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# File watching
watchdog>=3.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Code Quality
flake8>=7.0.0
black>=24.0.0
mypy>=1.8.0

# Type stubs (optional)
types-regex>=2023.0.0
types-pyyaml>=6.0

# Web UI (v1.0.3)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0

# MCP Server (v1.0.5)
mcp[cli]>=1.0.0
python-dotenv
